Moved config stuff in config.go
